+++ R/model-survey.R
@@ -25,11 +25,23 @@ zsurvey$methods(zelig = function(formula, data, ids = ~1, probs = NULL,
                                 by = NULL, bootstrap = FALSE) {
     .self$zelig.call <- match.call(expand.dots = TRUE)
 
+    warning("Not all features are available in Zelig Survey.\nConsider using surveyglm and setx directly.\nFor details see: <http://docs.zeligproject.org/articles/to_zelig.html>.",
+            call. = FALSE)
+
     recastString2Formula <- function(a) {
         if (is.character(a)) {
             a <- as.Formula(paste("~", a))
         }
         return(a)
+    }
+
+    extract_vector <- function(x, df = data) {
+        if ("formula" %in% class(x))
+            x <- as.character(x)[[2]]
+        if (is.character(x))
+            if (x %in% names(df))
+                x <- df[, x]
+        return(x)
     }
 
     checkLogical <- function(a, name = "") {
@@ -53,7 +65,10 @@ zsurvey$methods(zelig = function(formula, data, ids = ~1, probs = NULL,
     ## character arguments.
     ids <- recastString2Formula(ids)
     probs <- recastString2Formula(probs)
-    localWeights <- recastString2Formula(localWeights)
+    # Convert to vector from data frame as formula experssion for weights was
+    # not being passed
+    localWeights <- extract_vector(localWeights)
+    #localWeights <- recastString2Formula(localWeights)
     strata <- recastString2Formula(strata)
     fpc <- recastString2Formula(fpc)
     checkforerror <- checkLogical(nest, "nest")
@@ -76,10 +91,12 @@ zsurvey$methods(zelig = function(formula, data, ids = ~1, probs = NULL,
                                         bootstrap.average = bootstrap.average,
             scale = scale, rscales = rscales, fpctype = fpctype, fpc = fpc)
     }
+
     .self$model.call <- as.call(list(.self$fn,
                                 formula = .self$zelig.call$formula,
                                 design = design))  # fn will be set again by super, but initialized here for clarity
     .self$model.call$family <- call(.self$family, .self$link)
-    callSuper(formula = formula, data = data, weights = localWeights, ..., by = by,
-              bootstrap = bootstrap)
+
+    callSuper(formula = formula, data = data, weights = localWeights, ...,
+              by = by, bootstrap = bootstrap)
 })